Move TextDocumentStore into separate module
