+++ aardvark-node/src/lib.rs
@@ -1,2 +1,5 @@
-pub mod network;
-pub mod operation;
+mod network;
+mod operation;
+mod store;
+
+pub use network::Network;
